app/requirements.txt
# --- Core Framework ---
fastapi
uvicorn[standard]
pydantic
python-dotenv

# --- HTTP & Async ---
httpx
aiofiles

# --- Databases ---
sqlalchemy
psycopg2-binary
motor              # MongoDB async client

# --- Security & Crypto ---
cryptography
passlib[bcrypt]
python-jose[cryptography]

# --- AI / LLM ---
openai
anthropic
ollama

# --- Vector / Embeddings ---
faiss-cpu
numpy

# --- NLP / Language ---
langdetect
sentence-transformers

# --- Web Search & Scraping ---
beautifulsoup4
lxml

# --- File Parsing ---
pypdf
python-docx
pandas
pillow

# --- Audio (STT/TTS) ---
soundfile
librosa
elevenlabs

# --- Utilities ---
tenacity
orjson
rich
pydantic_settings